refactor: move main view and log view logic into one file for now
