--- src/components/views/log/mod.rs
@@ -1,9 +1,0 @@
-pub mod log_view;
-
-pub use super::log::log_view::LogView;
-
-mod filter;
-mod subfilter;
-mod time_stamp;
-
-use super::log::{filter::Filter, subfilter::SubFilter, time_stamp::TimeStamp};
