--- src/components/views/mod.rs
@@ -1,9 +1,0 @@
-pub mod controls;
-pub mod log;
-pub mod task_view;
-
-pub use super::views::{controls::Controls, log::*, task_view::TaskView};
-
-mod paginator;
-
-use paginator::Paginator;
